+++ tests/server.rs
@@ -1,24 +1,76 @@
-// use crate::test_data::RAW_MAGIT_DIFF_RUST;
-use crate::test_data::*;
-
 #[cfg(test)]
 mod tests {
-
-    use tower_lsp::LanguageServer;
     use diff_lsp::server::get_backends_map;
     use diff_lsp::server::DiffLsp;
     use diff_lsp::{DiffHeader, MagitDiff, Parsable};
     use expanduser::expanduser;
     use log::info;
+    use std::fs;
     use tower_lsp::lsp_types::*;
+    use tower_lsp::LanguageServer;
     // use super::*;
     use tower_lsp::LspService;
+
+    fn get_init_params() -> tower_lsp::lsp_types::InitializeParams {
+        #[allow(deprecated)] // root_path is deprecated but without it, code doesn't compile? :(
+        InitializeParams {
+            process_id: None,
+            root_path: None,
+            root_uri: None,
+            initialization_options: None,
+            capabilities: ClientCapabilities {
+                workspace: None,
+                text_document: {
+                    Some(TextDocumentClientCapabilities {
+                        hover: Some(HoverClientCapabilities::default()),
+                        references: Some(ReferenceClientCapabilities {
+                            dynamic_registration: Some(false),
+                        }),
+                        ..Default::default()
+                    })
+                },
+                window: None,
+                general: None,
+                experimental: None,
+            },
+            trace: None,
+            workspace_folders: None,
+            client_info: None,
+            locale: None,
+        }
+    }
+
+    pub fn get_open_params_rust(uri: Url) -> tower_lsp::lsp_types::DidOpenTextDocumentParams {
+        DidOpenTextDocumentParams {
+            text_document: TextDocumentItem {
+                uri: (uri),
+                // TODO: Figure out how to select based on
+                language_id: "rust".to_string(),
+                version: 1,
+                // TODO: Allow which text as param to helper?
+                text: fs::read_to_string("tests/data/rust_diff.magit_status").unwrap(),
+            },
+        }
+    }
+
+    #[allow(dead_code)]
+    pub fn get_open_params_go(uri: Url) -> tower_lsp::lsp_types::DidOpenTextDocumentParams {
+        DidOpenTextDocumentParams {
+            text_document: TextDocumentItem {
+                uri: (uri),
+                language_id: "go".to_string(),
+                version: 1,
+                text: fs::read_to_string("tests/data/go_diff.magit_status").unwrap(),
+            },
+        }
+    }
 
     //#[tokio::test]
     #[allow(dead_code)]
     async fn test_end_to_end_rust_analyzer() {
         // Note this test depends on the environment having rust-analyzer installed and on the path.
-        let diff = MagitDiff::parse(RAW_MAGIT_DIFF_RUST).unwrap();
+        let diff_text = fs::read_to_string("tests/data/rust_diff.magit_status").unwrap();
+        let diff = MagitDiff::parse(&diff_text).unwrap();
         let root: String = expanduser("~/diff-lsp").unwrap().display().to_string();
 
         assert_eq!(
@@ -46,19 +98,12 @@ mod tests {
             },
         };
 
-        let _init_res = service
-            .inner()
-            .initialize(test_data::get_init_params())
-            .await
-            .unwrap();
+        let _init_res = service.inner().initialize(get_init_params()).await.unwrap();
 
         info!("_init_res: {:?}", _init_res);
 
         service.inner().initialized(InitializedParams {}).await;
-        service
-            .inner()
-            .did_open(test_data::get_open_params_rust(url))
-            .await;
+        service.inner().did_open(get_open_params_rust(url)).await;
 
         let hover_result = service.inner().hover(hover_request).await.unwrap().unwrap();
         info!("{:?}", hover_result);
@@ -68,7 +113,8 @@ mod tests {
     #[allow(dead_code)]
     async fn test_end_to_end_gopls() {
         // Note this test depends on the environment having gopls installed and on the path.
-        let diff = MagitDiff::parse(test_data::RAW_MAGIT_DIFF_GO).unwrap();
+        let diff_text = fs::read_to_string("tests/data/go_diff.magit_status").unwrap();
+        let diff = MagitDiff::parse(&diff_text).unwrap();
         let root: String = expanduser("~/diff-lsp").unwrap().display().to_string();
         info!("Root is {:?}", root);
 
@@ -96,19 +142,12 @@ mod tests {
             },
         };
 
-        let _init_res = service
-            .inner()
-            .initialize(test_data::get_init_params())
-            .await
-            .unwrap();
+        let _init_res = service.inner().initialize(get_init_params()).await.unwrap();
 
         info!("_init_res: {:?}", _init_res);
 
         service.inner().initialized(InitializedParams {}).await;
-        service
-            .inner()
-            .did_open(test_data::get_open_params_go(url))
-            .await;
+        service.inner().did_open(get_open_params_go(url)).await;
 
         let hover_result = service.inner().hover(hover_request).await.unwrap().unwrap();
         info!("{:?}", hover_result);
@@ -116,7 +155,8 @@ mod tests {
     // TODO move to lib.rs but having trouble importing test_data there
     #[test]
     fn test_parse_go_magit_diff() {
-        let parsed_diff = MagitDiff::parse(test_data::RAW_MAGIT_DIFF_GO).unwrap();
+        let diff_text = fs::read_to_string("tests/data/go_diff.magit_status").unwrap();
+        let parsed_diff = MagitDiff::parse(&diff_text).unwrap();
         assert_eq!(
             parsed_diff.headers.get(&DiffHeader::Buffer),
             Some(&"lsp-example".to_string())
